very little refactor
